+++ app/gui/src/project-view/stores/suggestionDatabase/entry.ts
@@ -240,3 +240,8 @@ export function makeArgument(name: string, type: string = 'Any'): SuggestionEntr
     hasDefault: false,
   }
 }
+
+/** `true` if calling the function without providing a value for this argument will result in an error. */
+export function isRequiredArgument(info: SuggestionEntryArgument) {
+  return !!info.defaultValue?.startsWith('Missing_Argument.')
+}