+++ CHANGELOG.md
@@ -55,6 +55,7 @@
   be opened on OS X][11755].
 - [Fix some UI elements drawing on top of visualization toolbar dropdown
   menus][11768].
+- [Highlight missing required arguments][11803].
 - [Arrows in some drop-down buttons are now clearly visible][11800]
 
 [11151]: https://github.com/enso-org/enso/pull/11151
@@ -89,6 +90,7 @@
 [11753]: https://github.com/enso-org/enso/pull/11753
 [11761]: https://github.com/enso-org/enso/pull/11761
 [11768]: https://github.com/enso-org/enso/pull/11768
+[11803]: https://github.com/enso-org/enso/pull/11803
 [11800]: https://github.com/enso-org/enso/pull/11800
 
 #### Enso Standard Library